+++ src/main.c
@@ -283,6 +283,7 @@ static const struct t_subdirs_entry cachedir_subdirs[] = {
     {DIR_CACHE_COVER,         "Cover cache dir"},
     {DIR_CACHE_LYRICS,        "Lyrics cache dir"},
     {DIR_CACHE_MISC,          "Misc cache dir"},
+    {DIR_CACHE_THUMBS, "Thumbs cache dir"},
     {DIR_CACHE_WEBRADIODB,    "Webradiodb cache dir"},
     {NULL, NULL}
 };